safari-600.7-branch/LayoutTests/css3/viewport-percentage-lengths/viewport-percentage-lengths-resize-expected.txt

Test that viewport lengths and font sizes properly resize with the viewport.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS successfullyParsed is true

TEST COMPLETE
PASS innerWidth is 800
PASS innerHeight is 600
PASS getComputedStyle(test).fontSize is "30px"
PASS getComputedStyle(test).width is "400px"
PASS getComputedStyle(testpseudo, ':after').marginLeft is "120px"
PASS getComputedStyle(testpseudo, ':after').paddingRight is "200px"
PASS innerWidth is 900
PASS innerHeight is 600
PASS getComputedStyle(test).fontSize is "30px"
PASS getComputedStyle(test).width is "450px"
PASS getComputedStyle(testpseudo, ':after').marginLeft is "120px"
FAIL getComputedStyle(testpseudo, ':after').paddingRight should be 225px. Was 200px.
PASS innerWidth is 900
PASS innerHeight is 640
PASS getComputedStyle(test).fontSize is "32px"
PASS getComputedStyle(test).width is "450px"
FAIL getComputedStyle(testpseudo, ':after').marginLeft should be 128px. Was 120px.
FAIL getComputedStyle(testpseudo, ':after').paddingRight should be 225px. Was 200px.
PASS innerWidth is 500
PASS innerHeight is 640
PASS getComputedStyle(test).fontSize is "32px"
PASS getComputedStyle(test).width is "250px"
FAIL getComputedStyle(testpseudo, ':after').marginLeft should be 100px. Was 120px.
FAIL getComputedStyle(testpseudo, ':after').paddingRight should be 160px. Was 200px.
PASS innerWidth is 800
PASS innerHeight is 600
PASS getComputedStyle(test).fontSize is "30px"
PASS getComputedStyle(test).width is "400px"
PASS getComputedStyle(testpseudo, ':after').marginLeft is "120px"
PASS getComputedStyle(testpseudo, ':after').paddingRight is "200px"
